+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Maps2: Updated Amenities selection support [#5344](https://github.com/rokwire/illinois-app/issues/5344).
+### Added
+- Maps2: added analytics events [#5344](https://github.com/rokwire/illinois-app/issues/5344).
 
 ## [7.3.11] - 2025-10-13
 ### Added